# docs/openapi.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "School Shop Backend",
    "version": "0.1.0",
    "description": "Private Spring Boot business API. Cookie mutations require allowlisted Origin. JWT is always verified when supplied. Prices are integer VND. Expected version and idempotency are mandatory where documented."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/seller/products": {
      "get": {
        "operationId": "get_seller_products",
        "tags": [
          "seller"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductViewPage"
                }
              }
            }
          },
          "400": {
            "description": "Standard API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Standard API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Standard API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Standard API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Standard API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Standard API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Standard API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Standard API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Standard API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "maximum": 100000
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 20,
              "maximum": 100
            }
          }
        ]
      },
      "post": {
        "operationId": "post_seller_products",
        "tags": [
          "seller"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductView"
                }
              }
            }
          },
          "400": {
            "description": "Standard API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Standard API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Standard API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Standard API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Standard API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Standard API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Standard API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Standard API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Standard API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProduct"
              }
            }
          }
        }
      }
    },
...
    "/seller/products/{id}/stock-adjustments": {
      "post": {
        "operationId": "post_seller_products_id_stock_adjustments",
        "tags": [
          "seller"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockView"
                }
              }
            }
          },
          "400": {
            "description": "Standard API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Standard API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Standard API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Standard API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Standard API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Standard API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Standard API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Standard API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Standard API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 22,
              "maxLength": 128
            },
            "description": "Random UUID v4. Reuse on retry with exactly the same payload."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockInput"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "StockInput": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "stockOnHand": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "deltaOnHand": {
            "type": "integer",
            "format": "int32"
          },
          "reason": {
            "type": "string",
            "maxLength": 500
          },
          "expectedVersion": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "reason",
          "expectedVersion"
        ],
        "description": "Supply exactly one of absolute stockOnHand or deltaOnHand; expectedVersion is inventoryVersion."
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "details": {
            "oneOf": [
              {
                "type": "array",
                "items": {
                  "type": "object"
                }
              },
              {
                "type": "object"
              }
            ]
          }
        },
        "required": [
          "code",
          "message",
          "requestId",
          "timestamp"
        ]
      },
...
      "ProductView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "categoryName": {
            "type": "string"
          },
          "price": {
            "type": "integer",
            "minimum": 0,
            "maximum": 99999999999999
          },
          "imageAssetId": {
            "type": "string",
            "format": "uuid"
          },
          "imageUrl": {
            "type": "string"
          },
          "isSoldOut": {
            "type": "boolean"
          },
          "availableStock": {
            "type": "integer",
            "format": "int32"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string"
          },
          "ingredients": {
            "type": "string"
          },
          "allergens": {
            "type": "string"
          },
          "preservationInstructions": {
            "type": "string"
          },
          "stockOnHand": {
            "type": "integer",
            "format": "int32"
          },
          "stockReserved": {
            "type": "integer",
            "format": "int32"
          },
          "inventoryVersion": {
            "type": "integer",
            "format": "int64"
          }
        },
        "description": "Public view omits stockOnHand, stockReserved, inventoryVersion. Seller view includes them."
      },
...
      "StockView": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "stockOnHand": {
            "type": "integer",
            "format": "int32"
          },
          "stockReserved": {
            "type": "integer",
            "format": "int32"
          },
          "availableStock": {
            "type": "integer",
            "format": "int32"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "ProductViewPage": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductView"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "CreateProduct": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100
          },
          "slug": {
            "type": "string",
            "maxLength": 120
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "description": {
            "type": "string",
            "maxLength": 5000
          },
          "price": {
            "type": "integer",
            "minimum": 1,
            "maximum": 99999999999999
          },
          "imageAssetId": {
            "type": "string",
            "format": "uuid"
          },
          "ingredients": {
            "type": "string",
            "maxLength": 2000
          },
          "allergens": {
            "type": "string",
            "maxLength": 2000
          },
          "storageInstructions": {
            "type": "string",
            "maxLength": 2000
          },
          "initialStock": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "expectedVersion": {
            "type": "integer",
            "format": "int64"
          },
          "stockOnHand": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "preservationInstructions": {
            "type": "string",
            "maxLength": 2000
          }
        },
        "required": [
          "name",
          "categoryId",
          "price"
        ]
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "guestCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "guest_order"
      },
      "checkoutCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "checkout_session"
      }
    }
  }
}